Finn.py and SimpleGraphics.py
SimpleGraphics.py is a library used for Python Basic Shapes and colours. Do not change the library itself.
Finn.py is the code used to design my Finn Drawing. Lots of basic mathematical addition and subtraction are used to set the shapes and adapt them into the center of the screen.